requirements.txt
torch
torchvision
torchaudio
numpy
Pillow
tqdm